Call getMatchableLocales in importMessages
